TINY-9957: Refactor accordion key overrides to use MatchKeys (#8947)
* TINY-9957: Create DetailsDelete module stub and add MatchKeys overrides in DeleteBackspaceKeys

* TINY-9957: Refactor Accordion key backspace/delete overrides

- move code from DetailsElement module to DetailsDelete module
- update DetailsDeleteTest to accommodate for refactor

* TINY-9957: Improve getModifiedGranularity function

* TINY-9957: Fix Safari deletion overrides

* TINY-9957: Fix modified granularity deletions on Safari

* TINY-9957: Resolve Safari-specific failures in DetailsDeleteTest

- Previously the test did not test the Safari-specific overrides
